hackish fix to buildings not getting placed on board
Without knowing the real cause to why buildings were making it into the game.buildings array but not getting placed on the board, this commit fixes that.

The board now stores the game that encapsulates it as a reference. The building is only pushed into the game.buildings array once the board confirms that the user has successfully placed the building on the board.

I am still fascinated as to why this was happening. I have an inkling that it has something to do with mis-timed asynchronous setInterval() calls related to the game loop and board loop.