fix building placement bug
I originally turned off the need to select a building every time you wanted to build one. This meant that a building was only being instantiated once per building button click. So when trying to place something on the board, it would point to the same object in memory as the last one placed.

To patch this in the short term, I have uncommented the line that resets the game.selectedBuilding. This forces users to select a building (and hence instantiate it) every time they want to build.